# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: OneQR Business Growth Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /businesses/{id}/analytics:
    get:
      operationId: getAnalytics
      tags: [analytics]
      summary: Get business analytics
      security: [{ BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: period
          in: query
          schema:
            type: string
            enum: [7d, 30d, 90d]
            default: 7d
      responses:
        "200":
          description: Analytics data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsResponse"
  /analytics/track:
    post:
      operationId: trackEvent
      tags: [analytics]
      summary: Track an analytics event (public)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrackEventRequest"
      responses:
        "204":
          description: Event tracked
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: List all users (admin only)
      security: [{ BearerAuth: [] }]
      parameters:
        - name: plan
          in: query
          schema:
            type: string
            enum: [free, pro, business]
        - name: status
          in: query
          schema:
            type: string
            enum: [active, expired, cancelled]
        - name: expiresBeforeDays
          in: query
          schema:
            type: integer
            description: Users whose subscription expires within N days
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUser"
  /admin/users/{id}/suspend:
    post:
      operationId: adminSuspendUser
      tags: [admin]
      summary: Suspend a user account
      security: [{ BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: User suspended
  /admin/users/{id}/unsuspend:
    post:
      operationId: adminUnsuspendUser
      tags: [admin]
      summary: Unsuspend a user account
      security: [{ BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: User unsuspended
  /admin/subscriptions/expire:
    post:
      operationId: adminExpireSubscriptions
      tags: [admin]
      summary: Expire subscriptions that have passed their expiry date
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          description: Number of subscriptions expired
          content:
            application/json:
              schema:
                type: object
                properties:
                  expired:
                    type: integer
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    User:
      type: object
      properties:
        id: { type: string }
        email: { type: string }
        name: { type: string }
        avatarUrl: { type: string, nullable: true }
        subscriptionPlan: { type: string, enum: [free, pro, business] }
        subscriptionStatus: { type: string, enum: [active, expired, cancelled, trial] }
        subscriptionExpiresAt: { type: string, format: date-time, nullable: true }
        isAdmin: { type: boolean }
        createdAt: { type: string, format: date-time }
      required: [id, email, name, subscriptionPlan, subscriptionStatus, isAdmin, createdAt]
    AdminUser:
      allOf:
        - $ref: "#/components/schemas/User"
        - type: object
          properties:
            isSuspended: { type: boolean }
            lastSeenAt: { type: string, format: date-time, nullable: true }
...
    AnalyticsResponse:
      type: object
      properties:
        totalScans: { type: integer }
        totalClicks: { type: integer }
        totalReviewClicks: { type: integer }
        daily:
          type: array
          items:
            $ref: "#/components/schemas/DailyStats"
        topLinks:
          type: array
          items:
            $ref: "#/components/schemas/LinkStats"
      required: [totalScans, totalClicks, totalReviewClicks, daily, topLinks]
    DailyStats:
      type: object
      properties:
        date: { type: string }
        qrScans: { type: integer }
        profileViews: { type: integer }
        linkClicks: { type: integer }
      required: [date, qrScans, profileViews, linkClicks]
    LinkStats:
      type: object
      properties:
        platform: { type: string }
        clicks: { type: integer }
        percentage: { type: number }
      required: [platform, clicks, percentage]
    TrackEventRequest:
      type: object
      properties:
        businessId: { type: string }
        eventType: { type: string }
        targetId: { type: string }
        targetLabel: { type: string }
      required: [businessId, eventType]